First working damage activations test

+++ core_40k_logic/src/main/java/core/DamageCalculation/RollResult.java
@@ -7,7 +7,7 @@ public class RollResult {
 
 
 
-    public float averageAmountOfWounds;
-    public float averageAmountOfModelsSlain;
+    public double averageAmountOfWounds;
+    public double averageAmountOfModelsSlain;
 
 }
